TestDocker/requirements.txt
numpy>=1.21.0
torch>=1.13.0
torchvision>=0.14.0
tqdm>=4.60.0
Pillow>=8.0.0
scikit-image>=0.19.0
scipy>=1.7.0
noise>=1.2.0